- fixed the type of bitops.
    33ce36f4d79796c7865397598cfd839cbe1f6bac
fixed the pipe latency to 256.
    bee14243ce34e0aaddfdba7523423db14894ab9f

+++ alsa/drivers/serialmidi.c
@@ -91,7 +91,7 @@ typedef struct _snd_serialmidi {
 	char *sdev;			/* serial device name (e.g. /dev/ttyS0) */
 	unsigned int speed;		/* speed in bauds */
 	unsigned int adaptor;		/* see SERIAL_ADAPTOR_ */
-	unsigned int mode;		/* see SERIAL_MODE_* */
+	unsigned long mode;		/* see SERIAL_MODE_* */
 	unsigned int outs;		/* count of outputs */
 	unsigned char prev_status[SNDRV_SERIAL_MAX_OUTS];
 	snd_rawmidi_t *rmidi;		/* rawmidi device */
